refactor(provider): move http client to separate dir
