+++ ethers-providers/src/transports/http.rs
@@ -4,14 +4,14 @@ use crate::{provider::ProviderError, JsonRpcClient};
 use async_trait::async_trait;
 use reqwest::{Client, Error as ReqwestError};
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
 use std::{
-    fmt,
     str::FromStr,
     sync::atomic::{AtomicU64, Ordering},
 };
 use thiserror::Error;
 use url::Url;
+
+use super::common::{JsonRpcError, Request, Response};
 
 /// A low-level JSON-RPC Client over HTTP.
 ///
@@ -42,7 +42,7 @@ pub enum ClientError {
     ReqwestError(#[from] ReqwestError),
     #[error(transparent)]
     /// Thrown if the response could not be parsed
-    JsonRpcError(#[from] errors::JsonRpcError),
+    JsonRpcError(#[from] JsonRpcError),
 }
 
 impl From<ClientError> for ProviderError {
@@ -116,90 +116,5 @@ impl Clone for Provider {
             client: self.client.clone(),
             url: self.url.clone(),
         }
-    }
-}
-
-// leak private type w/o exposing it
-mod errors {
-    use super::*;
-
-    #[derive(Serialize, Deserialize, Debug, Clone, Error)]
-    /// A JSON-RPC 2.0 error
-    pub struct JsonRpcError {
-        /// The error code
-        pub code: i64,
-        /// The error message
-        pub message: String,
-        /// Additional data
-        pub data: Option<Value>,
-    }
-
-    impl fmt::Display for JsonRpcError {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            write!(
-                f,
-                "(code: {}, message: {}, data: {:?})",
-                self.code, self.message, self.data
-            )
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-/// A JSON-RPC request
-struct Request<'a, T> {
-    id: u64,
-    jsonrpc: &'a str,
-    method: &'a str,
-    params: T,
-}
-
-impl<'a, T> Request<'a, T> {
-    /// Creates a new JSON RPC request
-    fn new(id: u64, method: &'a str, params: T) -> Self {
-        Self {
-            id,
-            jsonrpc: "2.0",
-            method,
-            params,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-struct Response<T> {
-    id: u64,
-    jsonrpc: String,
-    #[serde(flatten)]
-    data: ResponseData<T>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(untagged)]
-enum ResponseData<R> {
-    Error { error: errors::JsonRpcError },
-    Success { result: R },
-}
-
-impl<R> ResponseData<R> {
-    /// Consume response and return value
-    fn into_result(self) -> Result<R, errors::JsonRpcError> {
-        match self {
-            ResponseData::Success { result } => Ok(result),
-            ResponseData::Error { error } => Err(error),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn response() {
-        let response: Response<u64> =
-            serde_json::from_str(r#"{"jsonrpc": "2.0", "result": 19, "id": 1}"#).unwrap();
-        assert_eq!(response.id, 1);
-        assert_eq!(response.data.into_result().unwrap(), 19);
     }
 }
